orders api added and completed

+++ src/main/java/com/masai/BigBasketReplica/Dto/ItemDto.java
@@ -6,7 +6,8 @@ import lombok.Data;
 public class ItemDto {
     String itemName;
     Integer itemId;
-    Integer unitPrice;
+    Double unitPrice;
     Integer itemQty;
-    Integer totalPrice;
+    Double totalPrice;
+
 }
